refactor DTOs and ajust response
